requirements.txt
opencv-python~=4.8.1.78
numpy~=1.26.2
matplotlib~=3.8.2
pytesseract~=0.3.10
Pillow~=10.1.0
exif~=1.6.0
google~=3.0.0
google-generativeai
python-dotenv~=1.0.0
langchain~=0.0.346
transformers~=4.35.2
einops
accelerate~=0.25.0
bitsandbytes
torch~=2.1.1
torchvision~=0.16.1
openai~=1.3.8
tensorflow
invoice2data~=0.4.5
pymongo~=4.6.1
keras~=2.15.0